% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmtt.R
\name{lmtt}
\alias{lmtt}
\title{My SLR function lmtt}
\usage{
lmtt(X, Y)
}
\arguments{
\item{X}{The predictor}

\item{Y}{The response variable}
}
\value{
The output similar to lm function in R
}
\description{
A simple linear regression function
}
\examples{
lmtt(c(1,2,3), c(0.9,1.8,3.1))
}
